Former-commit-id: f5a387b09477c839151cca2021775185dcd2b5d0

+++ Modules/TArc/Sources/TArc/Testing/Private/Selftests/ComboBoxTest.h
@@ -329,7 +329,7 @@ DAVA_TARC_TESTCLASS(ComboBoxTest)
         .WillOnce(Invoke(this, &ComboBoxTest::TestComboModuleStart));
 
         EXPECT_CALL(*this, IndexChanged(_))
-        .Times(comboTestData.size());
+        .Times(static_cast<int>(comboTestData.size()));
     }
 
     bool TestComplete(const DAVA::String& testName) const override
